% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_LRR_BAF_to_oncosnp_cnv.R
\name{AddLRRBAF2OncosnpCNV}
\alias{AddLRRBAF2OncosnpCNV}
\title{Add the LRR and BAF to each segment reported in a OncoSNP CNV file.}
\usage{
AddLRRBAF2OncosnpCNV(cnvDt, qcDt, probeDt)
}
\arguments{
\item{cnvDt}{The OncoSNP standard file (.cnvs) data.table loaded using the
loadOncosnpCnvFile() function.}

\item{qcDt}{The OncoSNP quality control (.qc) data.table loaded using the
LoadOncosnpQcFile() function.}

\item{probeDt}{The PennCNV probe data file data.table loaded using the
LoadPennCNVProbeData() function.}
}
\value{
A modified data.table of the cnvDt that contains extra columns for
				 LRR, LRRShifted, BAF, etc.
}
\description{
Add the LRR and BAF to each segment reported in a OncoSNP CNV file.
}

