% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_state_info.R
\name{get_state_info}
\alias{get_state_info}
\title{Get Oncosnp State Information}
\usage{
get_state_info(state.type = "CLL")
}
\arguments{
\item{state.type}{The "state" level.}
}
\value{
A list with names corresponding to various details "categories" of
  state information
}
\description{
Returns information about the Oncosnp states in a list structure. This is
dependent on the input state type list that was used in the Oncosnp run.
}

