% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_oncosnp_cnv_file}
\alias{load_oncosnp_cnv_file}
\title{Load the OncoSNP CNV file (.cnv).}
\usage{
load_oncosnp_cnv_file(cnvFile, version = c("1.3.0"))
}
\arguments{
\item{cnvFile}{OncoSNP CNV file}
}
\value{
A data.table of the standard output file (.cnvs) that has its columns
        renamed to be programmer-friendly to use
}
\description{
Load the OncoSNP CNV file (.cnv).
}

