% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_oncosnp_qc_file}
\alias{load_oncosnp_qc_file}
\title{Load the quality control files (.qc) generated by OncoSNP.}
\usage{
load_oncosnp_qc_file(qcFile)
}
\arguments{
\item{qcFile}{File generated that can be generated in the oncosnp module or
a single qc file generated by oncosnp directly.}
}
\value{
A data.frame of the quality control file that has its columns
        renamed to be programmer-friendly to use.
}
\description{
Load the quality control files (.qc) generated by OncoSNP.
}

