% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadPennCNVProbeData}
\alias{LoadPennCNVProbeData}
\title{Load the PennCNV probe data.}
\usage{
LoadPennCNVProbeData(probeFile)
}
\arguments{
\item{probeFile}{The PennCNV probe file}
}
\value{
A data.table of the probe file
}
\description{
Load the PennCNV probe data.
}

