% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oncosnputils.R
\docType{package}
\name{oncosnpUtils}
\alias{oncosnpUtils}
\alias{oncosnpUtils-package}
\title{oncosnputils: A package for post-processing OncoSNP results}
\description{
oncosnputils: A package for post-processing OncoSNP results
}

