% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_oncosnp_to_probe.R
\name{AddOncosnp2PennCNVProbe}
\alias{AddOncosnp2PennCNVProbe}
\title{Add the OncoSNP information to the PennCNV probe data.}
\usage{
AddOncosnp2PennCNVProbe(cnvDt, qcDt, probeDt, ploidyConfig = 1)
}
\arguments{
\item{cnvDt}{The OncoSNP standard file (.cnvs) data.table loaded using the
loadOncosnpCnvFile() function.}

\item{qcDt}{The OncoSNP quality control (.qc) data.table loaded using the
LoadOncosnpQcFile() function.}

\item{probeDt}{The PennCNV probe data file data.table loaded using the
paramDescriptionLoadPennCNVProbeData() function.}

\item{ploidyConfig}{The ploidy configuration from OncoSNP that you want to
to map the information to.}
}
\value{
A modified data.table of the probeDt that contains extra columns for
   shifted LRR and ranks from OncoSNP
}
\description{
Each probe will have it's LRR value shifted and assigned a rank
(or multiple) according to the output of OncoSNP
}
\examples{
AddOncosnp2PennCNVProbe(..., ploidyConfig = 1)
AddOncosnp2PennCNVProbe(..., ploidyConfig = 2)
}

